refactor: 공통되는 style 분리
