+++ frontend/src/components/CompetitionDetail/AfterCompetition.tsx
@@ -1,11 +1,10 @@
-import { css } from '@style/css';
-
 import { CompetitionInfo } from '@/apis/competitions';
 
 import ViewDashboardButton from '../Main/Buttons/ViewDashboardButton';
 import CompetitionDetailInfo from './CompetitionDetailInfo';
 import CompetitionMembersInfo from './CompetitionMembersInfo';
 import ProblemList from './ProblemList';
+import { buttonContainerStyle, statusTextContainerStyle, statusTextStyle } from './styles/styles';
 
 interface Props {
   competitionId: number;
@@ -30,31 +29,3 @@ export default function AfterCompetition({ competitionId, competition }: Props) 
     </div>
   );
 }
-
-const statusTextContainerStyle = css({
-  display: 'inline-flex',
-  padding: '4px 16px',
-  justifyContent: 'center',
-  alignItems: 'center',
-  gap: '10px',
-  borderRadius: '999px',
-  border: '1px solid var(--alert-error-default, #E23636)',
-  background: 'var(--alert-error-dark, #751919)',
-});
-
-const statusTextStyle = css({
-  color: '#FFF',
-  fontFamily: 'Inter',
-  fontSize: '12px',
-  fontStyle: 'normal',
-  fontWeight: '500',
-  lineHeight: 'normal',
-});
-
-const buttonContainerStyle = css({
-  display: 'flex',
-  width: '900px',
-  justifyContent: 'flex-end',
-  alignItems: 'center',
-  gap: '10px',
-});
