+++ frontend/src/components/CompetitionDetail/DuringCompetition.tsx
@@ -1,11 +1,10 @@
-import { css } from '@style/css';
-
 import { CompetitionInfo } from '@/apis/competitions';
 
 import ViewDashboardButton from '../Main/Buttons/ViewDashboardButton';
 import EnterCompetitionButton from './Buttons/EnterCompetitionButton';
 import CompetitionDetailInfo from './CompetitionDetailInfo';
 import CompetitionMembersInfo from './CompetitionMembersInfo';
+import { buttonContainerStyle, statusTextContainerStyle, statusTextStyle } from './styles/styles';
 
 interface Props {
   competitionId: number;
@@ -32,31 +31,3 @@ export default function DuringCompetition({ competitionId, competition, startsAt
     </div>
   );
 }
-
-const statusTextContainerStyle = css({
-  display: 'inline-flex',
-  padding: '4px 16px',
-  justifyContent: 'center',
-  alignItems: 'center',
-  gap: '10px',
-  borderRadius: '999px',
-  border: '1px solid var(--alert-info-default, #82DD55)',
-  background: 'var(--alert-success-dark, #355A23)',
-});
-
-const statusTextStyle = css({
-  color: '#FFF',
-  fontFamily: 'Inter',
-  fontSize: '12px',
-  fontStyle: 'normal',
-  fontWeight: '500',
-  lineHeight: 'normal',
-});
-
-const buttonContainerStyle = css({
-  display: 'flex',
-  width: '900px',
-  justifyContent: 'flex-end',
-  alignItems: 'center',
-  gap: '10px',
-});
